SDR-331 refactor: CQRS 형태의 패키지 구조 변경
