+++ be/src/main/java/com/jjikmuk/sikdorak/aws/api/AwsCommandApi.java
@@ -1,14 +1,14 @@
-package com.jjikmuk.sikdorak.aws.controller;
+package com.jjikmuk.sikdorak.aws.api;
 
-import com.jjikmuk.sikdorak.aws.service.GeneratePreSignedURLService;
-import com.jjikmuk.sikdorak.aws.service.ImageMetaDataService;
-import com.jjikmuk.sikdorak.aws.service.request.PreSignedUrlCreateRequest;
-import com.jjikmuk.sikdorak.aws.service.response.PreSignedUrlCreateResponse;
+import com.jjikmuk.sikdorak.aws.command.app.GeneratePreSignedURLService;
+import com.jjikmuk.sikdorak.aws.command.app.ImageMetaDataService;
+import com.jjikmuk.sikdorak.aws.command.app.request.PreSignedUrlCreateRequest;
+import com.jjikmuk.sikdorak.aws.command.app.response.PreSignedUrlCreateResponse;
 import com.jjikmuk.sikdorak.common.ResponseCodeAndMessages;
 import com.jjikmuk.sikdorak.common.aop.UserOnly;
 import com.jjikmuk.sikdorak.common.response.CommonResponseEntity;
-import com.jjikmuk.sikdorak.user.auth.api.LoginUser;
 import com.jjikmuk.sikdorak.user.auth.api.AuthenticatedUser;
+import com.jjikmuk.sikdorak.user.auth.api.LoginUser;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -19,7 +19,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/images/url")
-public class AwsS3Controller {
+public class AwsCommandApi {
 
 	private final GeneratePreSignedURLService generatePresignedURLService;
 	private final ImageMetaDataService imageMetaDataService;
