+++ be/src/main/java/com/jjikmuk/sikdorak/aws/command/app/ImageMetaDataService.java
@@ -1,0 +1,9 @@
+package com.jjikmuk.sikdorak.aws.command.app;
+
+import com.jjikmuk.sikdorak.aws.command.domain.ImageMetaData;
+import com.jjikmuk.sikdorak.user.auth.api.LoginUser;
+
+public interface ImageMetaDataService {
+
+	ImageMetaData initImageMetaData(String fileName, LoginUser loginUser);
+}
