+++ be/src/main/java/com/jjikmuk/sikdorak/aws/command/domain/ImageMetaData.java
@@ -1,4 +1,4 @@
-package com.jjikmuk.sikdorak.aws.domain;
+package com.jjikmuk.sikdorak.aws.command.domain;
 
 import com.jjikmuk.sikdorak.common.domain.BaseTimeEntity;
 import javax.persistence.Column;
