+++ be/src/main/java/com/jjikmuk/sikdorak/aws/exception/DuplicateImageMetaDataException.java
@@ -1,9 +1,9 @@
-package com.jjikmuk.sikdorak.aws.service.exception;
+package com.jjikmuk.sikdorak.aws.exception;
 
 import com.jjikmuk.sikdorak.common.exception.SikdorakRuntimeException;
 import org.springframework.http.HttpStatus;
 
-public class NotFoundImageMetaDataException extends SikdorakRuntimeException {
+public class DuplicateImageMetaDataException extends SikdorakRuntimeException {
 
 	@Override
 	public HttpStatus getHttpStatus() {
