+++ be/src/main/java/com/jjikmuk/sikdorak/aws/exception/NotFoundImageMetaDataException.java
@@ -1,9 +1,9 @@
-package com.jjikmuk.sikdorak.aws.service.exception;
+package com.jjikmuk.sikdorak.aws.exception;
 
 import com.jjikmuk.sikdorak.common.exception.SikdorakRuntimeException;
 import org.springframework.http.HttpStatus;
 
-public class DuplicateImageMetaDataException extends SikdorakRuntimeException {
+public class NotFoundImageMetaDataException extends SikdorakRuntimeException {
 
 	@Override
 	public HttpStatus getHttpStatus() {
