--- be/src/main/java/com/jjikmuk/sikdorak/aws/service/ImageMetaDataService.java
@@ -1,9 +1,0 @@
-package com.jjikmuk.sikdorak.aws.service;
-
-import com.jjikmuk.sikdorak.aws.domain.ImageMetaData;
-import com.jjikmuk.sikdorak.user.auth.controller.LoginUser;
-
-public interface ImageMetaDataService {
-
-	ImageMetaData initImageMetaData(String fileName, LoginUser loginUser);
-}
