+++ be/src/test/java/com/jjikmuk/sikdorak/unittest/aws/ImageExtensionTest.java
@@ -3,8 +3,8 @@ package com.jjikmuk.sikdorak.unittest.aws;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.jjikmuk.sikdorak.aws.domain.ImageExtension;
-import com.jjikmuk.sikdorak.aws.service.exception.InvalidImagesExtensionException;
+import com.jjikmuk.sikdorak.aws.command.domain.ImageExtension;
+import com.jjikmuk.sikdorak.aws.exception.InvalidImagesExtensionException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.params.ParameterizedTest;
